requirements.txt
Django==5.2.7
yt-dlp==2025.9.26
requests==2.32.5
python-telegram-bot==21.5
